added correct answer selection
inserting actual title of quiz in db instead of "Test"

quiz slide working, sending requests to quiz sesh on the server

+++ api.go
@@ -10,7 +10,9 @@ import (
 	_ "github.com/lib/pq"
 )
 
-// add more later
+// handleQuizGet qets the quizID from the end of the given URL w
+// and writes the info json from the db back using r.
+// ((add more later))
 func handleQuizGet(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	qID, err := strconv.Atoi(vars["id"])
@@ -18,26 +20,19 @@ func handleQuizGet(w http.ResponseWriter, r *http.Request) {
 		fmt.Println(err)
 	} else {
 		//fmt.Printf("%d", qID) //testing
-		rows, err := db.Query(`SELECT * FROM quiz WHERE qid=$1`, qID)
+		var info string
+		err := db.QueryRow(`SELECT info FROM quiz WHERE qid=$1`, qID).Scan(&info)
 		if err != nil {
 			fmt.Printf("%s", err)
 		} else {
-			for rows.Next() {
-				//fmt.Printf("here") //testing
-				var qid, cid int
-				var title, info string
-				err = rows.Scan(&qid, &title, &info, &cid)
-				if err != nil {
-					fmt.Printf("%s", err)
-				} else {
-					fmt.Printf("\nqid:%d \ttitle:%s \tinfo:%s \tcid:%d", qid, title, info, cid)
-				}
-			}
+			fmt.Fprintf(w, info)
 		}
 	}
 }
 
-// just an idea, not sure if we actually need this
+// handleQuizUpdate qets the quizID from the end of the given URL w,
+// gets the form data, and updates the quiz in the db.
+// ((just an idea, not sure if we actually need this))
 func handleQuizUpdate(w http.ResponseWriter, r *http.Request) {
 	fmt.Println("HERE")
 	vars := mux.Vars(r)
@@ -57,28 +52,29 @@ func handleQuizUpdate(w http.ResponseWriter, r *http.Request) {
 			fmt.Println("\nUpdated.")
 		}
 	}
-
 }
 
+// handleQuizList returns a json of all quiz ids and titles
+// using r.
 func handleQuizDelete(w http.ResponseWriter, r *http.Request) {
-  authenticate := auth(r)
-  if authentication != nil { 
-  vars := mux.Vars(r)
-  qid, err := strconv.Atoi(vars["id"])
-  if err != nil {
-    fmt.Println(err)
-  } else {
-    fmt.Println(qid)
-    _, err := db.Exec(`DELETE FROM quiz WHERE qid=$1`, qid)
-    if err != nil {
-      fmt.Println(err)
-    } else {
-      fmt.Println("\nDeleted.")
-    }
-  }
-  } else { //this is bad, but we can decide this later...
-    fmt.Println("Error - you cannot delete a quiz")
-  }
+	authenticate := auth(r)
+	if authenticate != nil {
+		vars := mux.Vars(r)
+		qid, err := strconv.Atoi(vars["id"])
+		if err != nil {
+			fmt.Println(err)
+		} else {
+			fmt.Println(qid)
+			_, err := db.Exec(`DELETE FROM quiz WHERE qid=$1`, qid)
+			if err != nil {
+				fmt.Println(err)
+			} else {
+				fmt.Println("\nDeleted.")
+			}
+		}
+	} else { //this is bad, but we can decide this later...
+		fmt.Println("Error - you cannot delete a quiz")
+	}
 }
 
 func handleQuizList(w http.ResponseWriter, r *http.Request) {
